requirements.txt
streamlit==1.36.0
langchain==0.3.0
langchain-cohere==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
numpy==1.26.4
pdfminer.six==20240706
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.36.0
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
pinecone==5.1.0
pinecone-client==5.0.1
chardet==5.2.0
streamlit-chat==0.1.1
unstructured==0.15.12